+++ container/fmu.c
@@ -14,7 +14,8 @@
  * FMI-importer supporting FMI-2.0 and FMI-3.0. FMUs are handled through fmu_t pointers.
  */
 
-fmu_status_t fmu_set_inputs(fmu_t* fmu) {
+
+fmu_status_t fmu_set_inputs(const fmu_t* fmu) {
     fmu_status_t status = FMU_STATUS_OK;
 
     const container_t* container = fmu->container;
@@ -42,7 +43,7 @@ fmu_status_t fmu_set_inputs(fmu_t* fmu) {
     SET_INPUT(booleans, Boolean);
     SET_INPUT(booleans1, Boolean1);
 
-    /* Need to add a (cast) to avod a warning */
+    /* strings: Need to add a (cast) to avod a warning */
     for (int i = 0; i < fmu_io->strings.in.nb; i += 1) {
         const unsigned int fmu_vr = fmu_io->strings.in.translations[i].fmu_vr;
         const unsigned int local_vr = fmu_io->strings.in.translations[i].vr;
@@ -50,13 +51,103 @@ fmu_status_t fmu_set_inputs(fmu_t* fmu) {
         if (status != FMU_STATUS_OK) \
             return status; \
     }
+
+    /* binaries: Need to add size parameter */
+    for (int i = 0; i < fmu_io->binaries.in.nb; i += 1) {
+        const unsigned int fmu_vr = fmu_io->binaries.in.translations[i].fmu_vr;
+        const unsigned int local_vr = fmu_io->binaries.in.translations[i].vr;
+        status = fmuSetBinary(fmu, &fmu_vr, 1, &container->binaries[local_vr].size,
+                              (const uint8_t *const*)&container->binaries[local_vr].data);
+        if (status != FMU_STATUS_OK) \
+            return status; \
+    }
+
+    SET_INPUT(clocks, Clock);
+
 #undef SET_INPUT
 
     return status;
 }
 
+static bool fmu_get_clock(const fmu_t *fmu, fmu_vr_t fmu_vr) {
+    bool value;
+    fmi3ValueReference vr = fmu_vr;
 
-fmu_status_t fmu_get_outputs(fmu_t* fmu) {
+    if (fmu->fmi_functions.version_3.fmi3GetClock(fmu->component, &vr, 1, &value) != fmi3OK)
+        value = false;
+
+    return value;
+}
+
+fmu_status_t fmu_set_clocked_inputs(const fmu_t* fmu) {
+    fmu_status_t status = FMU_STATUS_OK;
+
+    const container_t* container = fmu->container;
+    const fmu_io_t* fmu_io = &fmu->fmu_io;
+
+#define SET_CLOCKED_INPUT(variable, fmi_type) \
+    for (unsigned long i = 0; i < fmu_io->clocked_ ## variable .nb_in; i += 1) { \
+        fmu_clocked_port_t *clocked_port = &fmu_io->clocked_ ## variable .in[i]; \
+        if (fmu_get_clock(fmu, clocked_port->clock_fmu_vr)) { \
+            for(unsigned long j=0; j < clocked_port->translations_list.nb; j += 1) { \
+                const unsigned int fmu_vr = clocked_port->translations_list.translations[j].fmu_vr; \
+                const unsigned int local_vr = clocked_port->translations_list.translations[j].vr; \
+                status = fmuSet ## fmi_type (fmu, &fmu_vr, 1, &container-> variable [local_vr]); \
+                if (status != FMU_STATUS_OK) \
+                    return status; \
+            } \
+        } \
+    }
+
+    SET_CLOCKED_INPUT(reals64, Real64);
+    SET_CLOCKED_INPUT(reals32, Real32);
+    SET_CLOCKED_INPUT(integers8, Integer8);
+    SET_CLOCKED_INPUT(uintegers8, UInteger8);
+    SET_CLOCKED_INPUT(integers16, Integer16);
+    SET_CLOCKED_INPUT(uintegers16, UInteger16);
+    SET_CLOCKED_INPUT(integers32, Integer32);
+    SET_CLOCKED_INPUT(uintegers32, UInteger32);
+    SET_CLOCKED_INPUT(integers64, Integer64);
+    SET_CLOCKED_INPUT(uintegers64, UInteger64);
+    SET_CLOCKED_INPUT(booleans, Boolean);
+    SET_CLOCKED_INPUT(booleans1, Boolean1);
+
+    /* strings: Need to add a (cast) to avod a warning */
+    for (unsigned long i = 0; i < fmu_io->clocked_strings.nb_in; i += 1) {
+        fmu_clocked_port_t *clocked_port = &fmu_io->clocked_strings.in[i];
+        if (fmu_get_clock(fmu, clocked_port->clock_fmu_vr)) {
+            for(unsigned long j=0; j < clocked_port->translations_list.nb; j += 1) {
+                const unsigned int fmu_vr = clocked_port->translations_list.translations[j].fmu_vr;
+                const unsigned int local_vr = clocked_port->translations_list.translations[j].vr;
+                status = fmuSetString(fmu, &fmu_vr, 1, (const char *const*)&container->strings[local_vr]);
+                if (status != FMU_STATUS_OK) \
+                    return status; \
+            } \
+        } \
+    }
+
+    /* binaries: Need to add size parameter */
+    for (unsigned long i = 0; i < fmu_io->clocked_binaries.nb_in; i += 1) {
+        fmu_clocked_port_t *clocked_port = &fmu_io->clocked_binaries.in[i];
+        if (fmu_get_clock(fmu, clocked_port->clock_fmu_vr)) {
+            for(unsigned long j=0; j < clocked_port->translations_list.nb; j += 1) {
+                const unsigned int fmu_vr = clocked_port->translations_list.translations[j].fmu_vr;
+                const unsigned int local_vr = clocked_port->translations_list.translations[j].vr;
+                status = fmuSetBinary(fmu, &fmu_vr, 1, &container->binaries[local_vr].size,
+                        (const uint8_t *const*)&container->binaries[local_vr].data);
+                if (status != FMU_STATUS_OK)
+                    return status;
+            }
+        }
+    }
+
+#undef SET_CLOCKED_INPUT
+
+    return status;
+}
+
+
+fmu_status_t fmu_get_outputs(const fmu_t* fmu) {
     container_t* container = fmu->container;
     const fmu_io_t* fmu_io = &fmu->fmu_io;
     fmu_status_t status = FMU_STATUS_OK;
@@ -70,33 +161,140 @@ fmu_status_t fmu_get_outputs(fmu_t* fmu) {
             return status; \
     }
 
-GET_OUTPUT(reals64, Real64);
-GET_OUTPUT(reals32, Real32);
-GET_OUTPUT(integers8, Integer8);
-GET_OUTPUT(uintegers8, UInteger8);
-GET_OUTPUT(integers16, Integer16);
-GET_OUTPUT(uintegers16, UInteger16);
-GET_OUTPUT(integers32, Integer32);
-GET_OUTPUT(uintegers32, UInteger32);
-GET_OUTPUT(integers64, Integer64);
-GET_OUTPUT(uintegers64, UInteger64);
-GET_OUTPUT(booleans, Boolean);
-GET_OUTPUT(booleans1, Boolean1);
+    GET_OUTPUT(reals64, Real64);
+    GET_OUTPUT(reals32, Real32);
+    GET_OUTPUT(integers8, Integer8);
+    GET_OUTPUT(uintegers8, UInteger8);
+    GET_OUTPUT(integers16, Integer16);
+    GET_OUTPUT(uintegers16, UInteger16);
+    GET_OUTPUT(integers32, Integer32);
+    GET_OUTPUT(uintegers32, UInteger32);
+    GET_OUTPUT(integers64, Integer64);
+    GET_OUTPUT(uintegers64, UInteger64);
+    GET_OUTPUT(booleans, Boolean);
+    GET_OUTPUT(booleans1, Boolean1);
 
-for (size_t i = 0; i < fmu_io->strings.out.nb; i += 1) {
-        const fmu_vr_t fmu_vr = fmu_io->strings.out.translations[i].fmu_vr;
-        const fmu_vr_t local_vr = fmu_io->strings.out.translations[i].vr;
-        const char* str;
-        status = fmuGetString(fmu, &fmu_vr, 1, &str);
-        container->strings[local_vr] = strdup(str);
-        if (status != FMU_STATUS_OK) \
-            return status; \
-}
+    /* strings */
+    for (size_t i = 0; i < fmu_io->strings.out.nb; i += 1) {
+            const fmu_vr_t fmu_vr = fmu_io->strings.out.translations[i].fmu_vr;
+            const fmu_vr_t local_vr = fmu_io->strings.out.translations[i].vr;
+            const char* str;
+            status = fmuGetString(fmu, &fmu_vr, 1, &str);
+            if (status != FMU_STATUS_OK)
+                return status;
+            free(container->strings[local_vr]);
+            container->strings[local_vr] = strdup(str);
+    }
+
+    /* binaries */
+    for (size_t i = 0; i < fmu_io->binaries.out.nb; i += 1) {
+            const fmu_vr_t fmu_vr = fmu_io->binaries.out.translations[i].fmu_vr;
+            const fmu_vr_t local_vr = fmu_io->binaries.out.translations[i].vr;
+            const uint8_t *data;
+            size_t size;
+            status = fmuGetBinary(fmu, &fmu_vr, 1, &size, &data);
+            if (status != FMU_STATUS_OK)
+                return status;
+            if (container->binaries[local_vr].max_size < size) {
+                container->binaries[local_vr].data = realloc(container->binaries[local_vr].data, size);
+                if (! container->binaries[local_vr].data) {
+                    logger(LOGGER_ERROR, "Cannot allocate memory for get output.");
+                    return FMU_STATUS_ERROR;
+                }
+            }
+            container->binaries[local_vr].size = size;
+            memcpy(container->binaries[local_vr].data, data, size);
+    }
+
+    GET_OUTPUT(clocks, Clock);
+
 #undef GET_OUTPUT
 
     /* cast conversion between local variables */
     convert_proceed(fmu->container, fmu->conversions);
 
+    return status;
+}
+
+
+fmu_status_t fmu_get_clocked_outputs(const fmu_t* fmu) {
+    container_t* container = fmu->container;
+    const fmu_io_t* fmu_io = &fmu->fmu_io;
+    fmu_status_t status = FMU_STATUS_OK;
+
+#define GET_CLOCKED_OUTPUT(variable, fmi_type) \
+    for (unsigned long i = 0; i < fmu_io->clocked_ ## variable .nb_out; i += 1) { \
+        fmu_clocked_port_t *clocked_port = &fmu_io->clocked_ ## variable .out[i]; \
+        if (fmu_get_clock(fmu, clocked_port->clock_fmu_vr)) { \
+            for(unsigned long j=0; j < clocked_port->translations_list.nb; j += 1) { \
+                const unsigned int fmu_vr = clocked_port->translations_list.translations[j].fmu_vr; \
+                const unsigned int local_vr = clocked_port->translations_list.translations[j].vr; \
+                status = fmuGet ## fmi_type (fmu, &fmu_vr, 1, &container-> variable [local_vr]); \
+                if (status != FMU_STATUS_OK) \
+                    return status; \
+            } \
+        } \
+    }
+
+    GET_CLOCKED_OUTPUT(reals64, Real64);
+    GET_CLOCKED_OUTPUT(reals32, Real32);
+    GET_CLOCKED_OUTPUT(integers8, Integer8);
+    GET_CLOCKED_OUTPUT(uintegers8, UInteger8);
+    GET_CLOCKED_OUTPUT(integers16, Integer16);
+    GET_CLOCKED_OUTPUT(uintegers16, UInteger16);
+    GET_CLOCKED_OUTPUT(integers32, Integer32);
+    GET_CLOCKED_OUTPUT(uintegers32, UInteger32);
+    GET_CLOCKED_OUTPUT(integers64, Integer64);
+    GET_CLOCKED_OUTPUT(uintegers64, UInteger64);
+    GET_CLOCKED_OUTPUT(booleans, Boolean);
+    GET_CLOCKED_OUTPUT(booleans1, Boolean1);
+
+        /* strings: Need to add a (cast) to avod a warning */
+    for (unsigned long i = 0; i < fmu_io->clocked_strings.nb_out; i += 1) {
+        fmu_clocked_port_t *clocked_port = &fmu_io->clocked_strings.out[i];
+        if (fmu_get_clock(fmu, clocked_port->clock_fmu_vr)) {
+            for(unsigned long j=0; j < clocked_port->translations_list.nb; j += 1) {
+                const unsigned int fmu_vr = clocked_port->translations_list.translations[j].fmu_vr;
+                const unsigned int local_vr = clocked_port->translations_list.translations[j].vr;
+                const char* str;
+                status = fmuGetString(fmu, &fmu_vr, 1, &str);
+                if (status != FMU_STATUS_OK)
+                    return status;
+                free(container->strings[local_vr]);
+                container->strings[local_vr] = strdup(str);
+            } 
+        }
+    }
+
+    /* binaries: Need to add size parameter */
+    for (unsigned long i = 0; i < fmu_io->clocked_binaries.nb_out; i += 1) {
+        fmu_clocked_port_t *clocked_port = &fmu_io->clocked_binaries.out[i];
+        if (fmu_get_clock(fmu, clocked_port->clock_fmu_vr)) {
+            for(unsigned long j=0; j < clocked_port->translations_list.nb; j += 1) {
+                const unsigned int fmu_vr = clocked_port->translations_list.translations[j].fmu_vr;
+                const unsigned int local_vr = clocked_port->translations_list.translations[j].vr;
+                const uint8_t *data;
+                size_t size;
+                status = fmuGetBinary(fmu, &fmu_vr, 1, &size, &data);
+                if (status != FMU_STATUS_OK)
+                    return status;
+                if (container->binaries[local_vr].max_size < size) {
+                    container->binaries[local_vr].data = realloc(container->binaries[local_vr].data, size);
+                    if (! container->binaries[local_vr].data) {
+                        logger(LOGGER_ERROR, "Cannot allocate memory for get output.");
+                        return FMU_STATUS_ERROR;
+                    }
+                }
+                container->binaries[local_vr].size = size;
+                memcpy(container->binaries[local_vr].data, data, size);
+            }
+        }
+    }
+
+#undef GET_CLOCKED_OUTPUT
+
+    /* No need to convert: clocked variable are FMI-3 specific ! */
+    
     return status;
 }
 
@@ -352,7 +550,6 @@ int fmu_load_from_directory(container_t *container, int i, const char *directory
 
 
 void fmu_unload(fmu_t *fmu) {
-
     logger(LOGGER_DEBUG, "Unload FMU %s", fmu->name);
 
     /* Stop the thread */
@@ -394,6 +591,13 @@ void fmu_unload(fmu_t *fmu) {
     for (int i = 0; i < fmu->fmu_io.start_strings.nb; i += 1)
         free((char*)fmu->fmu_io.start_strings.start_values[i].value);
     FREE_FMU_DATA(strings);
+
+    free(fmu->fmu_io.binaries.in.translations);
+    free(fmu->fmu_io.binaries.out.translations);
+
+    free(fmu->fmu_io.clocks.in.translations);
+    free(fmu->fmu_io.clocks.out.translations);
+
 #undef FREE_FMU_DATA
 
     return;
@@ -472,12 +676,44 @@ GETTER_3   (uint32_t,      UInteger32,                  fmi3GetUInt32);
 GETTER_3   (int64_t,       Integer64,                   fmi3GetInt64);
 GETTER_3   (uint64_t,      UInteger64,                  fmi3GetUInt64);
 GETTER_2   (int,           Boolean,    fmi2GetBoolean)
-GETTER_3   (bool,          Boolean1,                    fmi3GetBoolean)
+GETTER_3   (bool,          Boolean1,                    fmi3GetBoolean);
 GETTER_BOTH(const char *,  String,     fmi2GetString,   fmi3GetString);
 
 #undef GETTER_BOTH
 #undef GETTER_3
 #undef GETTER_2
+
+fmu_status_t fmuGetBinary(const fmu_t *fmu, const fmu_vr_t vr[], size_t nvr, size_t size[], const uint8_t *value[]) {
+    fmu_status_t status = FMU_STATUS_OK;
+
+    if (fmu->fmi_version == FMU_2) {
+        logger(LOGGER_ERROR, "%s: fmuGetBinary not supported.", fmu->name);
+        status = FMU_STATUS_ERROR;
+    } else {
+        fmi3Status status3 = fmu->fmi_functions.version_3.fmi3GetBinary(fmu->component, vr, nvr, size, value, nvr);
+        if (status3 != fmi3OK) {
+            logger(LOGGER_ERROR, "%s: fmuGetBinary status=%d", fmu->name, status3);
+            status = FMU_STATUS_ERROR;
+        }
+    }
+    return status;\
+}
+
+fmu_status_t fmuGetClock(const fmu_t *fmu, const fmu_vr_t vr[], size_t nvr, bool value[]) {
+    fmu_status_t status = FMU_STATUS_OK;
+
+    if (fmu->fmi_version == FMU_2) {
+        logger(LOGGER_ERROR, "%s: fmuGetClock not supported.", fmu->name);
+        status = FMU_STATUS_ERROR;
+    } else {
+        fmi3Status status3 = fmu->fmi_functions.version_3.fmi3GetClock(fmu->component, vr, nvr, value);
+        if (status3 != fmi3OK) {
+            logger(LOGGER_ERROR, "%s: fmuGetClock status=%d", fmu->name, status3);
+            status = FMU_STATUS_ERROR;
+        }
+    }
+    return status;\
+}
 
 
 #define SETTER_BOTH(type, fmi_type, fmi2_function, fmi3_function) \
@@ -559,6 +795,67 @@ SETTER_BOTH(char * const,  String,     fmi2SetString,   fmi3SetString);
 #undef SETTER_2
 
 
+fmu_status_t fmuSetBinary(const fmu_t *fmu, const fmu_vr_t vr[], size_t nvr, const size_t size[], const uint8_t *const value[]) {
+    fmu_status_t status = FMU_STATUS_OK;
+
+    if (fmu->fmi_version == FMU_2) {
+        logger(LOGGER_ERROR, "%s: fmuSetBinary not supported.", fmu->name);
+        status = FMU_STATUS_ERROR;
+    } else {
+        fmi3Status status3 = fmu->fmi_functions.version_3.fmi3SetBinary(fmu->component, vr, nvr, size, value, nvr);
+        if (status3 != fmi3OK) {
+            logger(LOGGER_ERROR, "%s: fmuSetBinary status=%d", fmu->name, status3);
+            status = FMU_STATUS_ERROR;
+        }
+    }
+    return status;\
+}
+
+fmu_status_t fmuSetClock(const fmu_t *fmu, const fmu_vr_t vr[], size_t nvr, const bool value[]) {
+    fmu_status_t status = FMU_STATUS_OK;
+
+    if (fmu->fmi_version == FMU_2) {
+        logger(LOGGER_ERROR, "%s: fmuSetClock not supported.", fmu->name);
+        status = FMU_STATUS_ERROR;
+    } else {
+        fmi3Status status3 = fmu->fmi_functions.version_3.fmi3SetClock(fmu->component, vr, nvr, value);
+        if (status3 != fmi3OK) {
+            logger(LOGGER_ERROR, "%s: fmuSetClock status=%d", fmu->name, status3);
+            status = FMU_STATUS_ERROR;
+        }
+    }
+    return status;\
+}
+
+fmu_status_t fmuUpdateDiscreteStates(const fmu_t *fmu) {
+    fmi3Boolean discreteStatesNeedUpdate;
+    fmi3Boolean terminateSimulation;
+    fmi3Boolean nominalsOfContinuousStatesChanged;
+    fmi3Boolean valuesOfContinuousStatesChanged;
+    fmi3Boolean nextEventTimeDefined;
+    fmi3Float64 nextEventTime;
+    do {
+        logger(LOGGER_ERROR, "************************* fmi3UpdateDiscreteStates(fmu=%s)...", fmu->name);
+        fmi3Status status = fmu->fmi_functions.version_3.fmi3UpdateDiscreteStates(fmu->component,
+            &discreteStatesNeedUpdate,
+            &terminateSimulation,
+            &nominalsOfContinuousStatesChanged,
+            &valuesOfContinuousStatesChanged,
+            &nextEventTimeDefined,
+            &nextEventTime);
+        logger(LOGGER_ERROR, "-------> %s status=%d needupdate = %d terminate=%d %d %d next=%d time=%e", fmu->name, status, discreteStatesNeedUpdate, terminateSimulation, nominalsOfContinuousStatesChanged, valuesOfContinuousStatesChanged, nextEventTimeDefined, nextEventTime);
+        
+    } while(discreteStatesNeedUpdate);
+    fmu_get_outputs(fmu);
+    fmu->fmi_functions.version_3.fmi3EnterStepMode(fmu->component);
+
+    //logger(LOGGER_ERROR, "FMU '%s' requested event handling which is not supported.", fmu->name);
+    //status = FMU_STATUS_ERROR;
+
+    return FMU_STATUS_OK;
+}
+
+
 fmu_status_t fmuDoStep(const fmu_t *fmu, 
                        double currentCommunicationPoint, 
                        double communicationStepSize) {
@@ -596,26 +893,13 @@ fmu_status_t fmuDoStep(const fmu_t *fmu,
             status = FMU_STATUS_ERROR;
         }
         if (eventHandlingNeeded) {
-            fmi3Boolean discreteStatesNeedUpdate;
-            fmi3Boolean terminateSimulation;
-            fmi3Boolean nominalsOfContinuousStatesChanged;
-            fmi3Boolean valuesOfContinuousStatesChanged;
-            fmi3Boolean nextEventTimeDefined;
-            fmi3Float64 nextEventTime;
+            logger(LOGGER_ERROR, "fmu=%s: eventHandling is TRUE)...", fmu->name);
             fmu->fmi_functions.version_3.fmi3EnterEventMode(fmu->component);
-            fmi3Status status = fmu->fmi_functions.version_3.fmi3UpdateDiscreteStates(fmu->component,
-                &discreteStatesNeedUpdate,
-                &terminateSimulation,
-                &nominalsOfContinuousStatesChanged,
-                &valuesOfContinuousStatesChanged,
-                &nextEventTimeDefined,
-                &nextEventTime);
-            logger(LOGGER_ERROR, "%s %d %d %d %d %d %d %e", fmu->name, status, discreteStatesNeedUpdate, terminateSimulation, nominalsOfContinuousStatesChanged, valuesOfContinuousStatesChanged, nextEventTimeDefined, nextEventTime);
-            fmu->fmi_functions.version_3.fmi3EnterStepMode(fmu->component);
-
-            //logger(LOGGER_ERROR, "FMU '%s' requested event handling which is not supported.", fmu->name);
-            //status = FMU_STATUS_ERROR;
-        }
+            fmu_set_inputs(fmu);
+            fmuUpdateDiscreteStates(fmu);
+        } else
+            logger(LOGGER_ERROR, "fmu=%s: eventHandling is FALSE)...", fmu->name);
+            
         if ((status3 == fmi3OK) || (status3 == fmi3Warning))
             status = FMU_STATUS_OK;
     }
@@ -649,6 +933,7 @@ fmu_status_t fmuEnterInitializationMode(const fmu_t *fmu) {
 
 
 fmu_status_t fmuExitInitializationMode(const fmu_t *fmu) {
+    logger(LOGGER_ERROR, "fmuExitInitializationMode()");
     if (fmu->fmi_version == 2) {
         fmi2Status status2 = fmu->fmi_functions.version_2.fmi2ExitInitializationMode(fmu->component);
         if (status2 == fmi2OK)
@@ -657,6 +942,7 @@ fmu_status_t fmuExitInitializationMode(const fmu_t *fmu) {
             return FMU_STATUS_ERROR;
     } else {
         fmi3Status status3 = fmu->fmi_functions.version_3.fmi3ExitInitializationMode(fmu->component);
+        fmuUpdateDiscreteStates(fmu);
         if (status3 == fmi3OK)
             return FMU_STATUS_OK;
         else
@@ -679,7 +965,7 @@ fmu_status_t fmuSetupExperiment(const fmu_t *fmu) {
         /* No such API in FMI-3.0 */
         status = FMU_STATUS_OK;
     }
-    
+
     return status;
 }
 
@@ -706,7 +992,7 @@ fmu_status_t fmuInstantiateCoSimulation(fmu_t *fmu, const char *instanceName) {
             fmu->resource_dir,
             fmi3False,  /* visible */
             logger_get_debug(),
-            fmi3False, /* eventModeUsed */
+            fmi3True, /* eventModeUsed */
             fmi3False, /* earlyReturnAllowed */
             NULL, /* requiredIntermediateVariables[] */
             0, /*  nRequiredIntermediateVariables */